Change uint16_t for key codes

+++ keylogger.h
@@ -3,6 +3,7 @@
 
 #include <linux/input.h>
 #include <stdbool.h>
+#include <stdint.h>
 #include <stdio.h>
 
 #include "behavior_subject.h"
@@ -18,7 +19,7 @@ typedef struct {
     const char *shifted_and_capslocked;
 } KeyMap;
 
-const char *get_key_name(int key_code, bool shift_pressed, bool capslock_active);
+const char *get_key_name(uint16_t key_code, bool shift_pressed, bool capslock_active);
 char *strdup_alloc(const char *s);
 char *find_keyboard_device(const char *target_device_name);
 void log_key(FILE *fp, BehaviorSubject *subject, bool ctrl_pressed, bool meta_pressed, bool alt_pressed,
